fix a small mistake in basic_example
In the previous version, the layers were defined starting
from the lowest elevation of the first subsurface layer.
However, since the method generateStackedTIN calls the LaGriT
method stack/layers, the input layers are used as inner
boundaries between all layers.

In brief, this means that previously the top surface was not
included in the example.

This is fixed by appending a layer of 0.0 m offset to the
input array.