tinerator:fix stacking layers
In the previous version, providing `n` z-coordinates for the `layers`
variable in generateStackedTIN created `n-1` layers, because the
provided values are used as coordinates of the layer interfaces. In
our use cases, this is unexpected behavior. In order to fix this, we
now added a check in the `stackLayers` method in `generate_triplane`
to test if the first entry of `layers` is `0.0`. If not, an additional
value of `0.0` is added to the list, ensuring that stacking starts from
the provided top surface mesh.

The implementation is tested with `basic_example`, and creates the
expected 5 layers.

This addresses the closed pull request lanl/LaGriT#115