add argument to set the exponent for delineation
Although the argument `exponent` is defined in
`watershed_delineation`-method `watershedDelineation`, the calling
method in `dem_class` (`watershed_delineation`) did not have this
argument, so the user could not set an exponent at the top level. This
caused delineation methods such as `Holmgren` and `Freeman` to fail.

This commit gives an `exponent` argument to `dem_class`-method
`watershed_delineation` and sets this exponent as input to
`watershed_delineation`-method `watershedDelineation`.

The interactive part still does not get `exponent` as an argument.

fixes #149