spacing
Tried various markdown options to control spacing so it is not too much, and not too little. This version uses html <br> and &nbsp; to explicitly add lines and spaces where wanted. Note a long series of <br> lines at end of commands, this fools the skip to link to put the end of alphabet commands at the top of page instead of bottom of page.